sdm845-common: parts: avoid setting the same thermal prop
It appears that thermal engine service will get restarted whenever
vendor.thermal.vr_mode is written, no matter if the prop remains
the same or not. As a result, every app launch triggers a thermal
engine restart because the ThermalUtils in XiaomiParts writes to
the thermal prop unconditionally.

Avoid the unnecessary restart by keeping a copy of the current
thermal profile in ThermalUtils and only writing to the prop when
the profile is changed.

Test: boot, check logcat when launching apps with the same thermal
      profile and with a different one

Signed-off-by: Chenyang Zhong <zhongcy95@gmail.com>